requirements.txt
python-telegram-bot~=21.10
ffmpeg-python~=0.2.0